fix file icons in static build
